(feat) move base files and Coinbase Pro to new location
