+++ hummingbot/connector/exchange/coinbase_pro/coinbase_pro_order_book_tracker_entry.py
@@ -1,6 +1,6 @@
 from hummingbot.core.data_type.order_book import OrderBook
 from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
-from hummingbot.market.coinbase_pro.coinbase_pro_active_order_tracker import CoinbaseProActiveOrderTracker
+from hummingbot.connector.exchange.coinbase_pro.coinbase_pro_active_order_tracker import CoinbaseProActiveOrderTracker
 
 
 class CoinbaseProOrderBookTrackerEntry(OrderBookTrackerEntry):
