+++ test/integration/test_coinbase_pro_user_stream_tracker.py
@@ -20,9 +20,9 @@ from hummingbot.core.utils.async_utils import (
     safe_ensure_future,
     safe_gather,
 )
-from hummingbot.market.coinbase_pro.coinbase_pro_order_book_message import CoinbaseProOrderBookMessage
-from hummingbot.market.coinbase_pro.coinbase_pro_market import CoinbaseProMarket, CoinbaseProAuth
-from hummingbot.market.coinbase_pro.coinbase_pro_user_stream_tracker import CoinbaseProUserStreamTracker
+from hummingbot.connector.exchange.coinbase_pro.coinbase_pro_order_book_message import CoinbaseProOrderBookMessage
+from hummingbot.connector.exchange.coinbase_pro.coinbase_pro_market import CoinbaseProMarket, CoinbaseProAuth
+from hummingbot.connector.exchange.coinbase_pro.coinbase_pro_user_stream_tracker import CoinbaseProUserStreamTracker
 from hummingbot.market.market_base import OrderType
